mp: fix combat-message forwarding + non-actor monster-state redraw
Two co-op sync bugs found via the WAN/local playtest:

1. Message forwarding silently stopped once a player's message log filled.
   The capture diffed Messages::size() before/after each action, but that's
   a ring buffer capped at MESSAGE_LIMIT (default 255) — once full it stops
   growing, so the delta is always 0 and nothing forwards. Added a monotonic
   Messages::appended_total() counter (never decremented by ring eviction)
   and switched every MP watermark to it. Bumping the limit would only delay
   the bug; this removes it. Touches messages.cpp/.h (not upstream-hot) + MP
   files only.

2. Non-acting player saw killed monsters one turn late (corpse kept standing
   until their own next turn). State was applied correctly on receipt but the
   client only repaints inside the input loop, which an idle/waiting player
   doesn't enter. Added invalidate+redraw+refresh after draining a host-state
   batch (gated on recv_count>0), mirroring the vehicle_step handler.

+++ src/mp_gamestate.h
@@ -329,10 +329,11 @@ void set_last_host_action_label( const std::string &label );
 // Server only: capture avatar-generated messages since pre_msg and queue them
 // for the client with "You" → host character name substitution.
 // Call from do_turn() after each handle_action() when a remote player is connected.
-void host_capture_avatar_msgs( size_t pre_msg );
+// pre_msg is a Messages::appended_total() snapshot (monotonic), not size().
+void host_capture_avatar_msgs( unsigned long long pre_msg );
 // Capture ALL messages from vehmove() and queue them for the remote client as
 // their own messages (vehicle collision/status messages the driver should see).
-void host_capture_vehmove_msgs( size_t pre_msg );
+void host_capture_vehmove_msgs( unsigned long long pre_msg );
 
 // Per-tile vehicle position broadcast, called from map::vehmove after each
 // successful vehproceed step.  Sends a slim "vehicle_step" packet with just
